requirements.txt
boto
django-tastypie
hookbox
django-storages
importlib
django-ses
pysolr
django-haystack==1.1
mysql-python
pil
Werkzeug
markdown==2.0.3
suds